Canged data type of init_time @reservations

+++ reservations/app/database/entities.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, Date, Float, ForeignKey, Boolean
+from sqlalchemy import Column, String, Integer, Date, Float, ForeignKey, Boolean, BigInteger
 from sqlalchemy.orm import relationship
 from datetime import date
 import time
@@ -7,7 +7,7 @@ try:
     from database.base import Base, engine, Session
 except:
     from base import Base, engine, Session
-    
+
 class Reservation(Base):
     __tablename__ = 'reservation'
 
@@ -72,7 +72,7 @@ class Element(Base):
     information = Column(String(1000))
     url         = Column(String(100))
     date        = Column(Date)
-    init_time   = Column(Integer)
+    init_time   = Column(BigInteger)
     end_time    = Column(Integer)
     price       = Column(Float)
     reserved    = Column(Boolean)
